+++ packages/divoc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-divoc
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [d7105c0]
+  - @openfn/language-common@2.3.2
+
 ## 0.1.0
 
 DIVOC (Digital Infrastructure for Verifiable Open Credentialing) is an